+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog
 
 ## Version 0.1.0
+* added martian client for the IBM Quantum REST API
+